Fix tests to work after the move of Resource
